+++ frontend/src/auth/ForgotPassword.jsx
@@ -1,16 +1,13 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
-import { app } from "./firebaseConfig.js";
-import { useNavigate } from "react-router-dom";
-import './styles/style.css';
-import logo from './assets/LogoBCB1.png';
+import { app } from "../utils/firebaseConfig";
+import '../styles/style.css';
+import logo from '../assets/LogoBCB1.png';
 
 const LoginForm = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [error, setError] = useState(null);
-  const navigate = useNavigate(); // 
-
 
   const handleLogin = async (e) => {
     e.preventDefault();
@@ -20,7 +17,6 @@ const LoginForm = () => {
       const userCredential = await signInWithEmailAndPassword(auth, email, password);
       const user = userCredential.user;
       console.log("Usuario autenticado:", user);
-      navigate("/maincontent");
     } catch (error) {
       setError(error.message);
     }
@@ -28,14 +24,16 @@ const LoginForm = () => {
 
   return (
     <div className="login-container">
-        <div className="logo-container">
-      <img src={logo} alt="Logo" className="logo" />
+      <div className="logo-container">
+        <img src={logo} alt="Logo" className="logo" />
       </div>
       <div className="wrapper">
         <div className="form-box login">
           <div className="col-md-6">
-            <h2 className="text-center mb-4">Iniciar Sesión</h2>
+            <h2 className="text-center mb-4">Recupera tu Contraseña</h2>
             {error && <div className="alert alert-danger">{error}</div>}
+            <p className="mt-3 text-center" style={{ color: "#0a4d68" }}>
+              Ingresa tu correo para cambiar tu contraseña</p>
             <form onSubmit={handleLogin}>
               <div className="input-box">
                 <label htmlFor="email" className="form-label" ></label>
@@ -49,27 +47,9 @@ const LoginForm = () => {
                   required
                 />
               </div>
-              <div className="input-box">
-                <label htmlFor="password" className="form-label"></label>
-                <input
-                  type="password"
-                  className="form-control"
-                  id="password"
-                  placeholder="Contraseña"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  required
-                />
-              </div>
-              <div className="input-box left-align">
-                <a href="/forgotpassword" className="small">¿Olvidaste tu contraseña?</a>
-              </div>
               <div className="d-grid gap-2 mt-3">
-                <button type="submit" className="btn btn-primary">Iniciar Sesión</button>
+                <button type="submit" className="btn btn-primary">Recuperar Cuenta</button>
               </div>
-              <p className="mt-3 text-center" style={{ color: "#0a4d68" }}>
-  ¿No tienes cuenta? <a href="/register" className="register-link"><strong>Regístrate</strong></a>
-</p>
             </form>
           </div>
         </div>
