+++ frontend/src/auth/index.js
@@ -1,0 +1,5 @@
+export { default as ForgotPassword } from './ForgotPassword';
+export { default as LoginForm } from './LoginForm';
+export { default as RegisterForm } from './RegisterForm';
+export * from './index';
+export { default as setPhoto } from './setPhoto';
