+++ frontend/src/private/dashboard/components/index.js
@@ -1,0 +1,3 @@
+export { default as Feed } from './Feed';
+export { default as Navigation } from './Navigation';
+export { default as Suggestions } from './Suggestions';
