+++ frontend/src/private/dashboard/components/styles/feed.css
@@ -1,0 +1,74 @@
+/* feed.css */
+
+.feed {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+  }
+  
+  .post {
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    margin: 20px 0;
+    padding: 20px;
+    width: 100%;
+    max-width: 600px;
+    background-color: #fff;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .post-header {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+  }
+  
+  .profile-image {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    margin-right: 10px;
+  }
+  
+  h3 {
+    margin: 0;
+    font-size: 18px;
+    font-weight: bold;
+  }
+  
+  .post-content {
+    font-size: 16px;
+  }
+  
+  img {
+    max-width: 100%;
+    margin-top: 10px;
+  }
+  
+  /* Media queries for responsiveness */
+  @media (max-width: 768px) {
+    .post {
+      width: 90%;
+    }
+  }
+  
+  @media (max-width: 576px) {
+    .post {
+      padding: 10px;
+    }
+  
+    .profile-image {
+      width: 30px;
+      height: 30px;
+    }
+  
+    h3 {
+      font-size: 16px;
+    }
+  
+    .post-content {
+      font-size: 14px;
+    }
+  }
+
