+++ frontend/src/private/dashboard/components/styles/style.css
@@ -1,0 +1,248 @@
+body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+  background-color: #121413;
+  background-size: cover;
+  background-position: center;
+}
+
+.header {
+  position: fixed;
+  top: 0; /* Alineación en la parte superior */
+  left: 0;
+  color: #fff; /* Color del texto en el header */
+  padding: 1px 0; /* Espaciado interno en el header */
+  text-align: center; /* Centrar el contenido del header */
+  width: 100%; /* Ancho del header */
+}
+
+.logo-container {
+  display: flex;
+  justify-content: center;
+  margin-bottom: -18px; /* Ajustar la distancia del logo hacia abajo */
+}
+.logo {
+  width: 80%; /* Cambia este valor según tus necesidades */
+  display: block;
+  margin: 0 auto; /* Centrar horizontalmente */
+  margin-top: -100px; /* Ajustar la distancia superior */
+  margin-right: 0.9px; /* Ajustar la distancia derecha */
+}
+
+.wrapper {
+  position: relative;
+  width: 400px;
+  height: 410px;
+  background-color: #e8dfdf;
+  border: 2px solid rgba(222, 255, 255, 0.5);
+  border-radius: 29px;
+  backdrop-filter: blur(20px);
+  box-shadow: 0 0 3px rgba(255, 255, 255, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.wrapper .form-box {
+  width: 100%;
+  padding: 40px;
+}
+
+.register-wrapper {
+  height: 430px;
+}
+
+.photo-wrapper {
+  height: 590px;
+}
+
+.form-box h2 {
+  color: #0a4d68;
+  font-size: 32 + px;
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+.login-container h1 {
+  color: #0a4d68;
+  font-size: 32px;
+  text-align: center;
+  margin-top: -20px;
+}
+
+.register-photo {
+  font-size: 28px; /* Tamaño del título */
+  text-align: center; /* Centrar el texto */
+  margin-bottom: 15px; /* Espacio inferior */
+  color: #191616; /* Color del texto */
+}
+
+.left-align {
+  text-align: left;
+}
+
+.left-align a {
+  font-family: 'Segoe UI', sans-serif; /* Fuente */
+  font-size: 0.95em; /* Tamaño de fuente */
+  font-style: italic; /* Cursiva */
+}
+
+.input-box {
+  position: relative;
+  width: 100%;
+  margin: 15px 0; /* Ajusta el margen vertical entre los inputs */
+}
+
+.input-box label {
+  position: absolute;
+  top: 50%;
+  left: 5px;
+  transform: translateY(-50%);
+  font-size: 1em;
+  color: #707070; /* Cambia el color de la etiqueta */
+  font-weight: 500;
+  pointer-events: none;
+  transition: 0.5s;
+  transform-origin: left;
+  background-color: rgb(255, 255, 255); /* Agrega un fondo blanco detrás de la etiqueta */
+  padding: 0 5px; /* Ajusta el espaciado interno de la etiqueta */
+}
+
+.input-box input {
+  width: 300px;
+  height: 100%;
+  background-color: rgb(255, 255, 255); /* Cambia el color de fondo a blanco */
+  border: 1px solid #707070; /* Cambia el color del borde */
+  border-radius: 5px; /* Añade esquinas redondeadas */
+  outline: none;
+  color: #000000;
+  font-size: 1em;
+  font-weight: 500;
+  padding: 10px;
+  padding: 10px; /* Ajusta el espaciado interno del input */
+  transition: border-color 0.3s ease; /* Agrega una transición al borde */
+}
+
+.input-box input:focus {
+  border-color: #0a4d68; /* Cambia el borde al enfocar el input */
+}
+
+.btn {
+  width: 90%;
+  height: 40px;
+  background: #0a4d68;
+  border: none;
+  outline: none;
+  border-radius: 15px;
+  cursor: pointer;
+  font-size: 1em;
+  color: #fff;
+  font-weight: 500;
+  margin: 20px 0;
+}
+
+.login-container .register-text {
+  margin-top: -10px;
+  color: rgb(10, 77, 104);
+}
+
+.alert-message {
+  color: #0a4d68;
+  font-weight: bold;
+}
+
+.h2.register {
+  margin-top: -100px;
+  margin-right: 0.9px;
+}
+
+.profile-image-preview-box {
+  width: 100%;
+  height: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #e8dfdf;
+  border: 2px solid rgba(222, 255, 255, 0.5);
+  border-radius: 10px;
+  margin-bottom: 20px;
+}
+
+.profile-image-preview-box img {
+  max-width: 100%;
+  max-height: 100%;
+  object-fit: cover;
+  border-radius: 10px;
+}
+
+.input-box.label-centered label {
+  left: 50%;
+  transform: translateX(-50%);
+}
+
+.form-label {
+  font-size: 1em;
+  color: #707070;
+  font-weight: 500;
+  pointer-events: none;
+  transition: 0.5s;
+  transform-origin: left;
+  background-color: rgb(255, 255, 255);
+  padding: 0 5px;
+}
+
+.main-content {
+  display: flex;
+  justify-content: space-between;
+}
+
+/* Estilo para el contenedor de la barra de navegación */
+.navigation-container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 250px; /* Ajusta el ancho según tus necesidades */
+
+  padding: 20px;
+}
+
+/* Estilo para el logo */
+.logo {
+  margin-bottom: 20px; /* Agrega espacio entre el logo y los íconos de navegación */
+}
+
+/* Estilo para el saludo y el nombre de usuario */
+.user-info {
+  margin-bottom: 20px; /* Agrega espacio entre el saludo y los íconos de navegación */
+  font-size: 20px; /* Ajusta el tamaño del texto */
+}
+
+/* Estilo para los íconos y texto en la barra de navegación */
+.nav-links {
+  width: 100%;
+  list-style: none;
+  padding: 0;
+}
+
+.nav-links .list-group-item {
+  border: none;
+  padding: 0;
+  margin: 0;
+  text-align: left;
+}
+
+.nav-links .nav-link {
+  display: flex;
+  align-items: center; /* Centra verticalmente el contenido */
+  padding: 10px 0;
+  text-decoration: none;
+  color: #ffffff; /* O cualquier otro color de texto que desees */
+  font-size: 18px; /* Ajusta el tamaño de la fuente */
+}
+
+.nav-links .nav-link svg {
+  margin-right: 10px; /* Agrega espacio entre el ícono y el texto */
+  font-size: 24px; /* Ajusta el tamaño del ícono */
+}
