+++ frontend/src/private/dashboard/components/styles/suggestions.css
@@ -1,0 +1,70 @@
+/* suggestions.css */
+
+.suggestions {
+    margin-top: 40px;
+    padding: 20px;
+    background-color: #fff;
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .user-list {
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .user {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+  }
+  
+  .profile-image {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    margin-right: 10px;
+  }
+  
+  p {
+    margin: 0;
+    font-size: 16px;
+    font-weight: bold;
+  }
+  
+  button {
+    margin-left: auto;
+    background-color: #3897f0;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    padding: 8px 16px;
+    cursor: pointer;
+  }
+  
+  button:hover {
+    background-color: #357ae8;
+  }
+  
+  /* Media queries for responsiveness */
+  @media (max-width: 576px) {
+    .suggestions {
+      padding: 10px;
+    }
+  
+    .profile-image {
+      width: 30px;
+      height: 30px;
+    }
+  
+    p {
+      font-size: 14px;
+    }
+  
+    button {
+      font-size: 12px;
+      padding: 6px 12px;
+    }
+  }
+
