+++ frontend/src/private/dashboard/index.js
@@ -1,0 +1,2 @@
+export { default as MainContent } from './MainContent';
+export * from './components';
