+++ frontend/src/private/index.js
@@ -1,0 +1,3 @@
+export * from './dashboard/components';
+export * from './dashboard';
+export * from './index';
